// src/api/_generated/quizzesGameIoBackend.schemas.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * quizzes-game-io-backend
 * The Quizzes Game IO Backend API
 * OpenAPI spec version: 0.1
 */
export interface SignUpWithUsernameDto {
    username: string;
    password: string;
}
export interface AuthTokenDto {
    accessToken: string;
}
export interface SignInWithUsernameDto {
    username: string;
    password: string;
}
/**
 * error code
 */
export type SignUpWithUsernameControllerSignUpWithUsername400Code = 'COMMON.REQUEST_VALIDATION_ERROR';
export type SignUpWithUsernameControllerSignUpWithUsername400 = {
    /** http status code */
    statusCode?: number;
    /** error message */
    message?: string;
    /** error code */
    code?: SignUpWithUsernameControllerSignUpWithUsername400Code;
};
/**
 * error code
 */
export type SignUpWithUsernameControllerSignUpWithUsername409Code = 'ACCOUNT.USERNAME_ALREADY_OCCUPIED';
export type SignUpWithUsernameControllerSignUpWithUsername409 = {
    /** http status code */
    statusCode?: number;
    /** error message */
    message?: string;
    /** error code */
    code?: SignUpWithUsernameControllerSignUpWithUsername409Code;
};
/**
 * error code
 */
export type SignInWithUsernameControllerSignInWithUsername400Code = 'COMMON.REQUEST_VALIDATION_ERROR';
export type SignInWithUsernameControllerSignInWithUsername400 = {
    /** http status code */
    statusCode?: number;
    /** error message */
    message?: string;
    /** error code */
    code?: SignInWithUsernameControllerSignInWithUsername400Code;
};
/**
 * error code
 */
export type SignInWithUsernameControllerSignInWithUsername401Code = 'AUTH.SIGN_IN_INFO_NOT_MATCHED';
export type SignInWithUsernameControllerSignInWithUsername401 = {
    /** http status code */
    statusCode?: number;
    /** error message */
    message?: string;
    /** error code */
    code?: SignInWithUsernameControllerSignInWithUsername401Code;
};